// Angular/shared/models/candidate-profile-sitecore.models.ts
import { ProfileProgressSettings } from "./profile-pogress-settings";

export class CandidateProfileSitecoreModel {
    newAppointmentStatus: string;   
    newRequestMessage: string;  
    unlockProgramButtonText: string;
    backButtonText: string;
    viewScoresButtonText: string;    
    profileUpdMessage: string;
    personalHeader: string;
    loginHeader: string;
    graduateHeader: string;
    workHeader: string;
    degreesHeader: string;
    communicationHeader: string;
    viewOfficialScoresButtonText: string;
    orderScoreButtonText: string;
    seeAllAssessmentsLinkText: string;
    findMoreProgramsButtonText: string;
    seeAllProgramsLinkText: string;
    readButtonText: string;
    profileHeader: string;
    yourInformationHeader: string;
    profileInfoMessage: string;
    passwordChangeText: string;
    assessmentsButtonText: string;
    noAssessmentsButtonText: string;
    noAssessmentFoundTextPast: string;
    noAssessmentFoundTextFuture: string;
    schoolFinderUrl: string;
    changePasswordUrl: string;
    pvueProfileUrl: string;
    profilePageUrl: string;
    orderSchoolReportUrl: string;
    assessmentUrl: string;
    officialScoreReportsBaseURL: string;
    esrProductPurchaseURL: string;
    scoreReportURL: string;
    reinstateScoreURL: string;
    rescheduleExamURL: string;
    cancelExamURL: string;
    cancelScoreURL: string;
    serviceTimeout: number;
    viewESRMessage: string;
    profileProgressSettings: ProfileProgressSettings;
}